+++ DockDoor/Views/Hover Window/WindowPreview Supporting/WindowDismissalContainer.swift
